Call methods like `foo()` not `foo ()`
A method call like `debug ("asdf")` is technically evaluating the contents of
the parens first as a grouped expression, *then* passing the result as the
only parameter. This can backfire if you were to write `debug ("a", "b")`.

Calling methods like `foo "bar"` is nice, but can cause confusion like this.